# gh-103092: Add a mutex to make the random state of rotatingtree concurrent-safe