--- a/Misc/NEWS.d/next/Library/2024-02-12-11-42-48.gh-issue-103092.sGMKr0.rst
+++ b/Misc/NEWS.d/next/Library/2024-02-12-11-42-48.gh-issue-103092.sGMKr0.rst
@@ -1,0 +1,1 @@
+Isolate :mod:`_lsprof` (apply :pep:`687`).