--- a/Modules/rotatingtree.c
+++ b/Modules/rotatingtree.c
@@ -16,9 +16,7 @@
 
 static unsigned int random_value = 1;
 static unsigned int random_stream = 0;
-#define _zero 0
 static PyMutex random_mutex = {0};
-#undef _zero
 
 static int
 randombits(int bits)